RC 1.0

+++ Assets/RagonSDK/IRagonManager.cs
@@ -6,12 +6,12 @@ namespace Ragon.Client
   {
     void OnAuthorized(BitBuffer payload);
     void OnReady();
-    void OnEntityCreated(int entityId, ushort entityType, int ownerId, BitBuffer payload);
+    void OnEntityCreated(int entityId, ushort entityType, ushort ownerId, BitBuffer payload);
     void OnEntityDestroyed(int entityId, BitBuffer payload);
     void OnEntityState(int entityId, BitBuffer payload);
     void OnEntityProperty(int entityId, int property, BitBuffer payload);
-    void OnEntityEvent(int entityId, int evntCode, BitBuffer payload);
-    void OnEvent(uint evntCode, BitBuffer payload);
+    void OnEntityEvent(int entityId, ushort evntCode, BitBuffer payload);
+    void OnEvent(ushort evntCode, BitBuffer payload);
     void OnLevel(string sceneName);
     void OnConnected();
     void OnDisconnected();
